Downloads/system-builder-hub/backend/tests/requirements-test.txt
# Testing dependencies for SBH Meta-Builder v2
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# Coverage
coverage>=7.0.0

# HTTP testing
httpx>=0.24.0
requests-mock>=1.10.0

# Database testing
pytest-postgresql>=4.1.0
pytest-redis>=2.0.0

# UI testing
playwright>=1.35.0
pytest-playwright>=0.4.0

# Performance testing
locust>=2.15.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
